backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
PyPDF2
python-docx
google-genai
chromadb